backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pyserial==3.5
pydantic==2.5.2
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.0
bcrypt==4.0.1
websockets==11.0.3